refactor(copyable): rename to CopyablePre component for enhanced copyable text functionality
